models.py: Use select_related to optimize values queries
When we get a list of values, we want to get the metric, report,
and target objects as well, so use select_related(). This trades
off returning more data for each row for less queries.
(Potentially it might make sense to select_related only for report
and let individual queries be used for report and target.)